--- services/apps/alcs/src/providers/cdogs/cdogs.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CdogsService } from './cdogs.service';
-
-describe('CdogsService', () => {
-  let service: CdogsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [CdogsService],
-    }).compile();
-
-    service = module.get<CdogsService>(CdogsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
